refactor: DRY up showcase code
